# [Kernel][VIP] dispatch merge_attn_states cuda kernel, half&bf16