--- a/csrc/attention/merge_attn_states.cu
+++ b/csrc/attention/merge_attn_states.cu
@@ -1,35 +1,44 @@
 #include <optional>
-#include <cuda_runtime.h>
-#include <cmath>
 #include <torch/all.h>
 #include <ATen/cuda/CUDAContext.h>
 #include <c10/cuda/CUDAGuard.h>
+#include <algorithm>
+
+#include "attention_dtypes.h"
+#include "attention_utils.cuh"
+
+namespace vllm {
 
 // Implements section 2.2 of https://www.arxiv.org/pdf/2501.01005
 // can be used to combine partial attention results (in the split-KV case)
-// May loop over num heads for large NUM_TOKENS
-template <const bool LOOP_OVER_HEAD, const bool OUTPUT_LSE>
+template <typename scalar_t, vllm::Fp8KVCacheDataType KV_DTYPE,
+          bool LOOP_OVER_HEAD>
 __global__ void merge_attn_states_kernel(
-    float* output,      // [NUM_TOKENS, NUM_HEADS, HEAD_SIZE]
+    scalar_t* output,   // [NUM_TOKENS, NUM_HEADS, HEAD_SIZE]
     float* output_lse,  // [NUM_HEADS, NUM_TOKENS]
-    const float* __restrict__ prefix_output,  // [NUM_TOKENS, NUM_HEADS,
-                                              // HEAD_SIZE]
-    const float* __restrict__ prefix_lse,     // [NUM_HEADS, NUM_TOKENS]
-    const float* __restrict__ suffix_output,  // [NUM_TOKENS, NUM_HEADS,
-                                              // HEAD_SIZE]
-    const float* __restrict__ suffix_lse,     // [NUM_HEADS, NUM_TOKENS]
-    const uint NUM_TOKENS,                    // NUM_TOKENS
-    const uint NUM_HEADS,                     // NUM QUERY HEADS
-    const uint HEAD_SIZE                      // HEAD_SIZE, 32,48,64,...,512,etc
+    const scalar_t* __restrict__ prefix_output,  // [NUM_TOKENS, NUM_HEADS,
+                                                 // HEAD_SIZE]
+    const float* __restrict__ prefix_lse,        // [NUM_HEADS, NUM_TOKENS]
+    const scalar_t* __restrict__ suffix_output,  // [NUM_TOKENS, NUM_HEADS,
+                                                 // HEAD_SIZE]
+    const float* __restrict__ suffix_lse,        // [NUM_HEADS, NUM_TOKENS]
+    const uint num_tokens,                       // NUM_TOKENS
+    const uint num_heads,                        // NUM QUERY HEADS
+    const uint head_size  // HEAD_SIZE, 32,48,64,...,512,etc
 ) {
+  // TODO(DefTruth): may need to support fp8?
+  static_assert(KV_DTYPE == Fp8KVCacheDataType::kAuto,
+                "Currently merge_attn_states_kernel is not support for FP8.");
+
   if constexpr (LOOP_OVER_HEAD) {
+    // May loop over num heads for large NUM_TOKENS
     const uint token_idx = blockIdx.x;
     const uint thread_idx = threadIdx.x;
 
 #pragma unroll
-    for (uint head_idx = 0; head_idx < NUM_HEADS; ++head_idx) {
-      float p_lse = prefix_lse[head_idx * NUM_TOKENS + token_idx];
-      float s_lse = suffix_lse[head_idx * NUM_TOKENS + token_idx];
+    for (uint head_idx = 0; head_idx < num_heads; ++head_idx) {
+      float p_lse = prefix_lse[head_idx * num_tokens + token_idx];
+      float s_lse = suffix_lse[head_idx * num_tokens + token_idx];
       p_lse =
           std::isinf(p_lse) ? -std::numeric_limits<float>::infinity() : p_lse;
       s_lse =
@@ -42,40 +51,60 @@ __global__ void merge_attn_states_kernel(
       const float s_se = expf(s_lse);
       const float out_se = p_se + s_se;
 
-      if constexpr (OUTPUT_LSE) {
-        if (output_lse != nullptr) {
-          float out_lse = logf(out_se) + max_lse;
-          output_lse[head_idx * NUM_TOKENS + token_idx] = out_lse;
-        }
+      if (output_lse != nullptr) {
+        float out_lse = logf(out_se) + max_lse;
+        output_lse[head_idx * num_tokens + token_idx] = out_lse;
       }
 
       const uint blk_offset =
-          token_idx * NUM_HEADS * HEAD_SIZE + head_idx * HEAD_SIZE;
-      const uint thr_offset = thread_idx * 4;
+          token_idx * num_heads * head_size + head_idx * head_size;
+      const scalar_t* prefix_output_blk = prefix_output + blk_offset;
+      const scalar_t* suffix_output_blk = suffix_output + blk_offset;
+      scalar_t* output_blk = output + blk_offset;
+
+      // float -> 4, half/bf16 -> 8
+      using pack_128b_t = uint4;
+      constexpr uint pack_size = 16 / sizeof(scalar_t);
+
+      const uint thr_offset = thread_idx * pack_size;
       const float p_scale = p_se / out_se;
       const float s_scale = s_se / out_se;
-      const float* prefix_output_blk = prefix_output + blk_offset;
-      const float* suffix_output_blk = suffix_output + blk_offset;
-      float* output_blk = output + blk_offset;
 
-      if (thr_offset < HEAD_SIZE) {
-        float4 p_out4 = ((const float4*)(prefix_output_blk))[thr_offset / 4];
-        float4 s_out4 = ((const float4*)(suffix_output_blk))[thr_offset / 4];
+      if (thr_offset < head_size) {
+        // Pack 128b load
+        pack_128b_t p_out_pack = reinterpret_cast<const pack_128b_t*>(
+            prefix_output_blk)[thr_offset / pack_size];
+        pack_128b_t s_out_pack = reinterpret_cast<const pack_128b_t*>(
+            suffix_output_blk)[thr_offset / pack_size];
+        pack_128b_t o_out_pack;
 
-        float4 out4 = make_float4(p_out4.x * p_scale + s_out4.x * s_scale,
-                                  p_out4.y * p_scale + s_out4.y * s_scale,
-                                  p_out4.z * p_scale + s_out4.z * s_scale,
-                                  p_out4.w * p_scale + s_out4.w * s_scale);
-        ((float4*)(output_blk))[thr_offset / 4] = out4;
+#pragma unroll
+        for (uint i = 0; i < pack_size; ++i) {
+          // Always use float for FMA to keep precision.
+          // half(uint16_t), bfloat16, float -> float.
+          const float p_out_f =
+              vllm::to_float(reinterpret_cast<const scalar_t*>(&p_out_pack)[i]);
+          const float s_out_f =
+              vllm::to_float(reinterpret_cast<const scalar_t*>(&s_out_pack)[i]);
+          // fma: a * b + c = p_out_f * p_scale + (s_out_f * s_scale)
+          const float o_out_f = p_out_f * p_scale + (s_out_f * s_scale);
+          // float -> half(uint16_t), bfloat16, float.
+          vllm::from_float(reinterpret_cast<scalar_t*>(&o_out_pack)[i],
+                           o_out_f);
+        }
+
+        // Pack 128b storage
+        reinterpret_cast<pack_128b_t*>(output_blk)[thr_offset / pack_size] =
+            o_out_pack;
       }
-    }  // end loop over heads
+    }  // End loop over heads
   } else {
     const uint token_idx = blockIdx.x;
     const uint head_idx = blockIdx.y;
     const uint thread_idx = threadIdx.x;
 
-    float p_lse = prefix_lse[head_idx * NUM_TOKENS + token_idx];
-    float s_lse = suffix_lse[head_idx * NUM_TOKENS + token_idx];
+    float p_lse = prefix_lse[head_idx * num_tokens + token_idx];
+    float s_lse = suffix_lse[head_idx * num_tokens + token_idx];
     p_lse = std::isinf(p_lse) ? -std::numeric_limits<float>::infinity() : p_lse;
     s_lse = std::isinf(s_lse) ? -std::numeric_limits<float>::infinity() : s_lse;
 
@@ -86,43 +115,124 @@ __global__ void merge_attn_states_kernel(
     const float s_se = expf(s_lse);
     const float out_se = p_se + s_se;
 
-    if constexpr (OUTPUT_LSE) {
-      if (output_lse != nullptr) {
-        float out_lse = logf(out_se) + max_lse;
-        output_lse[head_idx * NUM_TOKENS + token_idx] = out_lse;
-      }
+    if (output_lse != nullptr) {
+      float out_lse = logf(out_se) + max_lse;
+      output_lse[head_idx * num_tokens + token_idx] = out_lse;
     }
 
     const uint blk_offset =
-        token_idx * NUM_HEADS * HEAD_SIZE + head_idx * HEAD_SIZE;
-    const uint thr_offset = thread_idx * 4;
+        token_idx * num_heads * head_size + head_idx * head_size;
+    const scalar_t* prefix_output_blk = prefix_output + blk_offset;
+    const scalar_t* suffix_output_blk = suffix_output + blk_offset;
+    scalar_t* output_blk = output + blk_offset;
+
+    // float -> 4, half/bf16 -> 8
+    using pack_128b_t = uint4;  // 16 bytes
+    constexpr uint pack_size = 16 / sizeof(scalar_t);
+
+    const uint thr_offset = thread_idx * pack_size;
     const float p_scale = p_se / out_se;
     const float s_scale = s_se / out_se;
-    const float* prefix_output_blk = prefix_output + blk_offset;
-    const float* suffix_output_blk = suffix_output + blk_offset;
-    float* output_blk = output + blk_offset;
 
-    if (thr_offset < HEAD_SIZE) {
-      float4 p_out4 = ((const float4*)(prefix_output_blk))[thr_offset / 4];
-      float4 s_out4 = ((const float4*)(suffix_output_blk))[thr_offset / 4];
+    if (thr_offset < head_size) {
+      // Pack 128b load
+      pack_128b_t p_out_pack = reinterpret_cast<const pack_128b_t*>(
+          prefix_output_blk)[thr_offset / pack_size];
+      pack_128b_t s_out_pack = reinterpret_cast<const pack_128b_t*>(
+          suffix_output_blk)[thr_offset / pack_size];
+      pack_128b_t o_out_pack;
 
-      float4 out4 = make_float4(p_out4.x * p_scale + s_out4.x * s_scale,
-                                p_out4.y * p_scale + s_out4.y * s_scale,
-                                p_out4.z * p_scale + s_out4.z * s_scale,
-                                p_out4.w * p_scale + s_out4.w * s_scale);
+#pragma unroll
+      for (uint i = 0; i < pack_size; ++i) {
+        // Always use float for FMA to keep precision.
+        // half(uint16_t), bfloat16, float -> float.
+        const float p_out_f =
+            vllm::to_float(reinterpret_cast<const scalar_t*>(&p_out_pack)[i]);
+        const float s_out_f =
+            vllm::to_float(reinterpret_cast<const scalar_t*>(&s_out_pack)[i]);
+        // fma: a * b + c = p_out_f * p_scale + (s_out_f * s_scale)
+        const float o_out_f = p_out_f * p_scale + (s_out_f * s_scale);
+        // float -> half(uint16_t), bfloat16, float.
+        vllm::from_float(reinterpret_cast<scalar_t*>(&o_out_pack)[i], o_out_f);
+      }
 
-      ((float4*)(output_blk))[thr_offset / 4] = out4;
+      // Pack 128b storage
+      reinterpret_cast<pack_128b_t*>(output_blk)[thr_offset / pack_size] =
+          o_out_pack;
     }
   }
 }
 
-#define LAUNCH_MERGE_ATTN_STATES_KERNEL(LOOP_OVER_HEAD, OUTPUT_LSE)        \
-  {                                                                        \
-    merge_attn_states_kernel<LOOP_OVER_HEAD, OUTPUT_LSE><<<grid, block>>>( \
-        output.data_ptr<float>(), output_lse_ptr,                          \
-        prefix_output.data_ptr<float>(), prefix_lse.data_ptr<float>(),     \
-        suffix_output.data_ptr<float>(), suffix_lse.data_ptr<float>(),     \
-        NUM_TOKENS, NUM_HEADS, HEAD_SIZE);                                 \
+}  // namespace vllm
+
+// The following macro is used to dispatch the conversion function based on
+// the output data type. The FN is a macro that calls a function with
+// template<typename SCALAR_T, KV_DTYPE>.
+#define DISPATCH_BY_SCALAR_DTYPE(SCALAR_DTYPE, FN)                      \
+  {                                                                     \
+    if (SCALAR_DTYPE == at::ScalarType::Float) {                        \
+      FN(float, vllm::Fp8KVCacheDataType::kAuto);                       \
+    } else if (SCALAR_DTYPE == at::ScalarType::Half) {                  \
+      FN(uint16_t, vllm::Fp8KVCacheDataType::kAuto);                    \
+    } else if (SCALAR_DTYPE == at::ScalarType::BFloat16) {              \
+      FN(__nv_bfloat16, vllm::Fp8KVCacheDataType::kAuto);               \
+    } else {                                                            \
+      TORCH_CHECK(false, "Unsupported data type of O: ", SCALAR_DTYPE); \
+    }                                                                   \
+  }
+
+#define LAUNCH_MERGE_ATTN_STATES(SCALAR_T, KV_DTYPE, LOOP_OVER_HEAD)        \
+  {                                                                         \
+    vllm::merge_attn_states_kernel<SCALAR_T, KV_DTYPE, LOOP_OVER_HEAD>      \
+        <<<grid, block>>>(                                                  \
+            reinterpret_cast<SCALAR_T*>(output.data_ptr()), output_lse_ptr, \
+            reinterpret_cast<SCALAR_T*>(prefix_output.data_ptr()),          \
+            reinterpret_cast<float*>(prefix_lse.data_ptr()),                \
+            reinterpret_cast<SCALAR_T*>(suffix_output.data_ptr()),          \
+            reinterpret_cast<float*>(suffix_lse.data_ptr()), num_tokens,    \
+            num_heads, head_size);                                          \
+  }
+
+template <typename SCALAR_T, vllm::Fp8KVCacheDataType KV_DTYPE>
+void merge_attn_states_launcher(
+    torch::Tensor& output,  // [NUM_TOKENS, NUM_HEADS, HEAD_SIZE]
+    std::optional<torch::Tensor> output_lse,  // [NUM_HEADS, NUM_TOKENS]
+    const torch::Tensor& prefix_output,  // [NUM_TOKENS, NUM_HEADS, HEAD_SIZE]
+    const torch::Tensor& prefix_lse,     // [NUM_HEADS, NUM_TOKENS]
+    const torch::Tensor& suffix_output,  // [NUM_TOKENS, NUM_HEADS, HEAD_SIZE]
+    const torch::Tensor& suffix_lse,     // [NUM_HEADS, NUM_TOKENS]
+    const bool disable_loop_over_head) {
+  const uint num_tokens = output.size(0);
+  const uint num_heads = output.size(1);  // num query heads
+  const uint head_size = output.size(2);
+  // float -> 4, half/bf16 -> 8, 128b = 16 bytes.
+  constexpr uint pack_size = 16 / sizeof(SCALAR_T);
+  TORCH_CHECK(head_size % pack_size == 0,
+              "headsize must be multiple of pack_size:", pack_size);
+  TORCH_CHECK(head_size / pack_size <= 1024,
+              "headsize/pack_size must be <= of 1024, pack_size: ", pack_size);
+  float* output_lse_ptr = nullptr;
+  if (output_lse.has_value()) {
+    output_lse_ptr = output_lse.value().data_ptr<float>();
+  }
+
+  if (num_tokens <= 1024 || num_heads >= 64 || disable_loop_over_head) {
+    dim3 grid(num_tokens, num_heads);
+    dim3 block(head_size / pack_size);
+    LAUNCH_MERGE_ATTN_STATES(SCALAR_T, KV_DTYPE, false);
+  } else {
+    // try loop over num heads for large NUM_TOKENS
+    dim3 grid(num_tokens);
+    dim3 block(head_size / pack_size);
+    LAUNCH_MERGE_ATTN_STATES(SCALAR_T, KV_DTYPE, true);
+  }
+}
+
+#define CALL_MERGE_ATTN_STATES_LAUNCHER(SCALAR_T, KV_DTYPE)           \
+  {                                                                   \
+    merge_attn_states_launcher<SCALAR_T, KV_DTYPE>(                   \
+        output, output_lse, prefix_output, prefix_lse, suffix_output, \
+        suffix_lse, disable_loop_over_head);                          \
   }
 
 void merge_attn_states(
@@ -133,32 +243,5 @@ void merge_attn_states(
     const torch::Tensor& suffix_output,  // [NUM_TOKENS, NUM_HEADS, HEAD_SIZE]
     const torch::Tensor& suffix_lse,     // [NUM_HEADS, NUM_TOKENS]
     const bool disable_loop_over_head) {
-  const uint NUM_TOKENS = output.size(0);
-  const uint NUM_HEADS = output.size(1);  // num query heads
-  const uint HEAD_SIZE = output.size(2);
-  assert(HEAD_SIZE % 4 == 0);     // headsize must be multiple of 4
-  assert(HEAD_SIZE / 4 <= 1024);  // headsize must be <= of 4096
-  float* output_lse_ptr = nullptr;
-  if (output_lse.has_value()) {
-    output_lse_ptr = output_lse.value().data_ptr<float>();
-  }
-
-  if (NUM_TOKENS <= 1024 || NUM_HEADS >= 64 || disable_loop_over_head) {
-    dim3 grid(NUM_TOKENS, NUM_HEADS);
-    dim3 block(HEAD_SIZE / 4);
-    if (output_lse_ptr != nullptr) {
-      LAUNCH_MERGE_ATTN_STATES_KERNEL(false, true);
-    } else {
-      LAUNCH_MERGE_ATTN_STATES_KERNEL(false, false);
-    }
-  } else {
-    // try loop over num heads for large NUM_TOKENS
-    dim3 grid(NUM_TOKENS);
-    dim3 block(HEAD_SIZE / 4);
-    if (output_lse_ptr != nullptr) {
-      LAUNCH_MERGE_ATTN_STATES_KERNEL(true, true);
-    } else {
-      LAUNCH_MERGE_ATTN_STATES_KERNEL(true, false);
-    }
-  }
+  DISPATCH_BY_SCALAR_DTYPE(output.dtype(), CALL_MERGE_ATTN_STATES_LAUNCHER);
 }