--- a/vllm/attention/ops/merge_attn_states.py
+++ b/vllm/attention/ops/merge_attn_states.py
@@ -3,6 +3,7 @@ from typing import Optional
 
 import torch
 
+import vllm.envs as envs
 from vllm.logger import init_logger
 from vllm.platforms import current_platform
 
@@ -20,7 +21,9 @@ def merge_attn_states(
     # NOTE(DefTruth): Currently, custom merge_attn_states CUDA kernel
     # is not support for FP8 dtype, fallback to use Triton kernel.
     supported_dtypes = [torch.float32, torch.half, torch.bfloat16]
-    if current_platform.is_cuda() and (output.dtype in supported_dtypes):
+    if ((not envs.VLLM_DISABLE_MERGE_ATTN_CUDA_OP)
+            and current_platform.is_cuda()
+            and (output.dtype in supported_dtypes)):
         from vllm._custom_ops import merge_attn_states
         return merge_attn_states(output, prefix_output, prefix_lse,
                                  suffix_output, suffix_lse, output_lse)